refactor(client): rehome async delegates, commit metadata resolvers, and streaming metadata handler
Tidy the org.apache.hudi.client root down to genuinely client-scoped classes:
- BaseCompactor, BaseClusterer -> org.apache.hudi.async: their only consumers
  are the async table services; they are the async delegation contract.
- CommitMetadataResolver, CommitMetadataResolverFactory,
  MarkerBasedCommitMetadataResolver, NoOpCommitMetadataResolver ->
  org.apache.hudi.table.action.commit, beside BaseCommitActionExecutor which
  drives them. The factory is a pure internal version/engine switch - not a
  user plug point.
- StreamingMetadataWriteHandler -> org.apache.hudi.metadata, joining the
  HoodieTableMetadataWriter it orchestrates; getMetadataWriter widens to
  protected for the cross-package engine subclasses.

Clients, client mixins, the frozen WriteStatus family and its stat value
objects, and CommitMetadataProperties stay in org.apache.hudi.client.

Behavior-preserving; no moved class is @PublicAPIClass, config-FQN-referenced,
or Kryo-registered.
