+++ hudi-client/hudi-client-common/src/main/java/org/apache/hudi/async/BaseClusterer.java
@@ -17,7 +17,9 @@
  * under the License.
  */
 
-package org.apache.hudi.client;
+package org.apache.hudi.async;
+
+import org.apache.hudi.client.BaseHoodieWriteClient;
 
 import java.io.IOException;
 import java.io.Serializable;
