+++ hudi-client/hudi-client-common/src/main/java/org/apache/hudi/metadata/StreamingMetadataWriteHandler.java
@@ -16,8 +16,9 @@
  * limitations under the License.
  */
 
-package org.apache.hudi.client;
+package org.apache.hudi.metadata;
 
+import org.apache.hudi.client.WriteStatus;
 import org.apache.hudi.common.data.HoodieData;
 import org.apache.hudi.common.model.HoodieCommitMetadata;
 import org.apache.hudi.common.model.HoodieWriteStat;
@@ -26,7 +27,6 @@ import org.apache.hudi.common.util.Option;
 import org.apache.hudi.common.util.ValidationUtils;
 import org.apache.hudi.common.util.VisibleForTesting;
 import org.apache.hudi.exception.HoodieException;
-import org.apache.hudi.metadata.HoodieTableMetadataWriter;
 import org.apache.hudi.table.HoodieTable;
 
 import java.util.HashMap;
@@ -89,7 +89,7 @@ public abstract class StreamingMetadataWriteHandler {
    * @return The metadata writer option.
    */
   @VisibleForTesting
-  synchronized Option<HoodieTableMetadataWriter> getMetadataWriter(String triggeringInstant, HoodieTable table) {
+  protected synchronized Option<HoodieTableMetadataWriter> getMetadataWriter(String triggeringInstant, HoodieTable table) {
 
     if (!table.getMetaClient().getTableConfig().getTableVersion().greaterThanOrEquals(HoodieTableVersion.EIGHT)) {
       return Option.empty();
