+++ hudi-client/hudi-spark-client/src/main/java/org/apache/hudi/async/SparkAsyncClusteringService.java
@@ -19,7 +19,6 @@
 
 package org.apache.hudi.async;
 
-import org.apache.hudi.client.BaseClusterer;
 import org.apache.hudi.client.BaseHoodieWriteClient;
 import org.apache.hudi.client.HoodieSparkClusteringClient;
 import org.apache.hudi.common.engine.HoodieEngineContext;
